moving around dtype conversion

+++ src/diffusers/pipelines/stable_video_diffusion/pipeline_stable_video_diffusion.py
@@ -120,8 +120,8 @@ class StableVideoDiffusionPipeline(DiffusionPipeline):
         self.text_encoder = self.text_encoder.to(device)
         if not isinstance(text, torch.Tensor):
             text = self.tokenizer(text, padding=True, return_tensors="pt")
-        text = text.to(device=device, dtype=dtype)
-        image_embeddings = self.text_encoder(**inputs).text_embeds.unsqueeze(1)
+        text = text.to(device=device)
+        image_embeddings = self.text_encoder(**inputs).text_embeds.unsqueeze(1).to(dtype)
 
         # duplicate image embeddings for each generation per prompt, using mps friendly method
         bs_embed, seq_len, _ = image_embeddings.shape
